pyproject.toml
[tool.poetry]
name = "geodude"
version = "0.1.0"
description = ""
authors = ["Andrew Li <andrewli@u.northwestern.edu>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.84.0"
pyyaml = "^6.0.2"
dotenv = "^0.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
